change all routes promises chaining to async await functions

+++ server/server.js
@@ -14,65 +14,63 @@ const port = process.env.PORT;
 
 app.use(bodyParser.json());
 
-app.post('/todos', authenticate, (req,res)=>{
-  new Todo({
-    title: req.body.title,
-    _creator: req.user._id
-  })
-    .save().
-    then((doc)=>res.send(doc))
-    .catch((e)=>res.status(400).send(e));
+app.post('/todos', authenticate, async (req,res)=>{
+  try {
+    const todo = await new Todo({title: req.body.title, _creator: req.user._id }).save();
+    res.send(todo);
+  } catch (e) {
+    res.status(400).send(e);
+  }
 });
 
 
-app.get('/todos',authenticate, (req, res)=>{
-  Todo.find({ _creator: req.user._id}).then((todos)=>{
-    res.send({todos});
-  }).catch((e)=>res.status(400).send(e));
+app.get('/todos',authenticate, async (req, res)=>{
+  try {
+    const todos = await Todo.find({ _creator: req.user._id });
+    res.send({ todos });
+  } catch (e) {
+    res.status(400).send(e)
+  }
 });
 
 
-app.get('/todos/:id',authenticate, (req, res)=>{
+app.get('/todos/:id',authenticate, async (req, res)=>{
+  const id = req.params.id;
+  if (!ObjectID.isValid(id)) {
+    return res.status(404).send({invalid:true});
+  }
+  try {
+    const todo = await Todo.findOne({_id: id,_creator: req.user._id});
+    if (!todo) {
+      return res.status(404).send({ empty: true });
+    }
+    res.send({ todo });
+  } catch (e) {
+    res.status(400).send(e)
+  }
+});
+
+
+app.delete('/todos/:id', authenticate, async (req, res)=>{
   var id = req.params.id;
   if (!ObjectID.isValid(id)) {
     return res.status(404).send({invalid:true});
   }
-
-  Todo.findOne({
-    _id: id,
-     _creator: req.user._id
-  }).then((todo)=>{
+  try {
+    const todo = await Todo.findOneAndRemove({ _id: id, _creator: req.user._id});
     if (!todo) {
-      return res.status(404).send({empty:true});
+      return res.status(404).send({ empty: true });
     }
-     res.send({todo});
-  }).catch((e)=>res.status(400).send(e));
-
-});
-
-
-app.delete('/todos/:id', authenticate, (req, res)=>{
-  var id = req.params.id;
-  if (!ObjectID.isValid(id)) {
-    return res.status(404).send({invalid:true});
+    res.send({ todo });
+  } catch (e) {
+    res.status(400).send(e)
   }
-
-  Todo.findOneAndRemove({
-    _id: id,
-    _creator: req.user._id
-  }).then((todo)=>{
-    if (!todo) {
-      return res.status(404).send({empty:true});
-    }
-     res.send({todo});
-  }).catch((e)=>res.status(400).send(e));
-
 });
 
 
-app.patch('/todos/:id',authenticate, (req, res)=>{
-  var id = req.params.id;
-  var body = _.pick(req.body, ['title','completed']);
+app.patch('/todos/:id',authenticate, async (req, res)=>{
+  const id = req.params.id;
+  const body = _.pick(req.body, ['title','completed']);
 
   if (!ObjectID.isValid(id)) {
     return res.status(404).send({invalid:true});
@@ -85,26 +83,29 @@ app.patch('/todos/:id',authenticate, (req, res)=>{
     body.completed = false;
   }
 
-  Todo.findOneAndUpdate({
-    _id: id,_creator: req.user._id},
-    {$set:body},{new:true}).then((todo)=>{
+  try {
+    const todo = await Todo.findOneAndUpdate({_id: id, _creator: req.user._id},
+                { $set: body }, { new: true });
     if (!todo) {
-      return res.status(404).send({empty:true});
+      return res.status(404).send({ empty: true });
     }
-     res.send({todo});
-  }).catch((e)=>res.status(400).send(e));
-
+    res.send({ todo });
+  } catch (e) {
+    res.status(400).send(e)
+  }
 });
 
 
-app.post('/users', (req, res) => {
-  var body = _.pick(req.body, ['email', 'password']);
-  var user = new User(body);
-    user.save().then(() => {
-      return user.generateAuthToken();
-    }).then((token) => {
-      res.header('x-auth', token).send(user);
-    }).catch((e) => res.status(400).send(e));
+app.post('/users', async (req, res) => {
+  try {
+    const body = _.pick(req.body, ['email', 'password']);
+    var user = new User(body);
+    await user.save();
+    const token = await user.generateAuthToken();
+    res.header('x-auth', token).send(user);
+  } catch (e) {
+    res.status(400).send(e)
+  }
 });
 
 
@@ -112,19 +113,24 @@ app.get('/users/me',authenticate ,(req, res) => {
   res.send(req.user);
 });
 
-app.post('/users/login', (req, res) => {
-  var body = _.pick(req.body, ['email', 'password']);
-  User.findByCredentials(body.email, body.password).then((user)=>{
-    return user.generateAuthToken().then((token)=>{
-      res.header('x-auth', token).send(user);
-    });
-  }).catch((e)=>res.status(400).send());
+app.post('/users/login', async (req, res) => {
+  try {
+    const body = _.pick(req.body, ['email', 'password']);    
+    const user = await User.findByCredentials(body.email, body.password);
+    const token = await user.generateAuthToken();
+    res.header('x-auth', token).send(user);
+  } catch (e) {
+    res.status(400).send()
+  }
 });
 
-app.delete('/users/me/token', authenticate, (req, res)=>{
-  req.user.removeToken(req.token).then(()=>{
-    res.status(200).send();
-  }).catch((e)=>res.status(400).send(e));
+app.delete('/users/me/token', authenticate, async (req, res)=>{
+  try {
+    await req.user.removeToken(req.token);
+    res.status(200).send();    
+  } catch (e) {
+    res.status(400).send(e)
+  }
 });
 
 app.listen(port, ()=>console.log('App running on port '+port));
